Movement and penguin checking
-Added basic movement functions called MovePeng* where * = NE, E, SE,
SW, W, NW. Will move only the penguin with given ID.
-Created CheckPeng for checking if there's a penguin on the given
coordinates. Return 1 if the space is free and 0 if there's a penguin on
given coordinates.
-PosPeng works now for multiple penguins, where number of penguins is
equal to PengRows. Will check if there's already a penguin in the same
spot.
-Updated DrawBoard to print out the multiple penguins.
-Added TURNBYTURN_MODE to main.h to give the option to clear screen
every DrawScreen() and wait for a button press from the user. This makes
turns easier to follow, but there's still the option to print out
everything (Windows only currently).